Kill patches/2.4.x and patches/3.2.14
The former is for an antiquated version of OTRS we haven't been using
for over a year. That's what git is for.

The latter... well, they're the right files but the *unpatched*
versions, which makes this kind of useless. No point in actually
correcting this, as this is also for a version of OTRS we're about to
move out from.

Change-Id: Ic6ac18e82594d5b7a122d2644d44704a8fae1fc6