Reverse-package our two OPMs into Source OPMs
The "opm" file format is OTRS' packaging format. It contains metadata,
as well as files base64-encoded in XML attributes. We've been checking
these out into git and modifying them as such, but this isn't actually
the way the format is designed: there is actually an "sopm" source
package format as well, which has the file contents separately.

The .opm can be reconstructed using something as simple as:
/opt/otrs/bin/otrs.PackageManager.pl -a build -p pkg/pkg.sopm -d pkg

Since the sources for our packages are probably lost in time, explode
our OPMs manually and reconstruct the original source packages. While at
it, move them from the 3.2-specific subdirectory as a package can
support multiple versions.

Change-Id: Ib77d6c982f5c88f6a03e64d32baa102c608b8f36