Move WikimediaTemplates' files under Custom/
OTRS actually supports local overrides with Custom/ (which we use in our
other package) and that was verified to work with templates as well, so
install our templates there instead. This actually allows our package to
be uninstalled without breaking the install, as well as easier diffs
between changed versions.

Change-Id: If509d12ece3b47de184b3f4d022d0ece2e47a976