Release 1.0.16. First version that only support 6.0.x
Unfortunately it doesn't seem possible or worthwhile to create a package
that supports both 5.0.x and 6.0.x as some of files have gone through
significant changes. Especially the AgentTicketZoom template seems to
have been split off into 2 separate templates (the second being a
subtemplate of sorts).

This patch was created by importing the new files as is and then reapply
the diff on top of them manually. In the process, we no longer patch
AgentTicketZoom, but rather AgentTicketZoom/TicketInformation.tt.

Bug: T187984
Change-Id: I127133df8e0402ba054c42b39d2ce49f977324fc